Simplify start of README.md
Simplify the introduction, particularly for new users that are likely to just want to click on a link to the user manual.

The description _LibreCAD's documentation repository_ is misleading because documentation includes the Website and Blog, Github and SourceForge repositories, User Manual, Wikis, Forum, Zulip chat Forum and IRC #librecad. A more accurate description might be _LibreCAD's Manuals_ (allowing for a future Programmers Manual).